+++ CHANGELOG.md
@@ -7,7 +7,10 @@ Changelog
 - Dropped Bignum support. Automatic conversion from Fixnum to Bignum (or vice versa)
   will produce columns whose values are of heterogeneous types, that are impossible to be read in a consistent way.
   You can use BigDecimal type instead for large numbers.
-- Added `HBase::Scope#while` which allows early termination of scan
+- Added `HBase::Scoped#while` which allows early termination of scan
   with [WhileMatchFilter](http://hbase.apache.org/apidocs/org/apache/hadoop/hbase/filter/WhileMatchFilter.html)
 - Filtering with regular expressions
-
+- Implemented comparator for `HBase::Result`
+- Added coprocessor administration methods
+- Basic aggregation with Coprocessor
+- `HBase::Scoped#count` with block